docs: add README.md file
This file introduces and explains a project. It contains information that is
commonly required to understand what the project is about. It's an easy way to
answer questions that your audience will likely have regarding how to install
and use your project and also how to collaborate with you.

This file is omitted when repository is downloaded.
More info: [Make a README](https://www.makeareadme.com).